/know/search -> /know/discover
